cmd/solrays: move IsValidHostname to pkg/util
